+++ services/src/main/java/org/keycloak/broker/oidc/OIDCIdentityProviderConfig.java
@@ -36,6 +36,8 @@ public class OIDCIdentityProviderConfig extends OAuth2IdentityProviderConfig {
     public static final String SUPPORTS_CLIENT_ASSERTION_REUSE = "supportsClientAssertionReuse";
     public static final String VALIDATE_REFRESH_TOKEN = "validateRefreshToken";
     public static final String CLAIMS_PARAMETER_SUPPORTED = "claimsParameterSupported";
+    public static final String ISSUER = "issuer";
+    public static final String LOGOUT_URL = "logoutUrl";
 
     public OIDCIdentityProviderConfig(IdentityProviderModel identityProviderModel) {
         super(identityProviderModel);
@@ -59,10 +61,10 @@ public class OIDCIdentityProviderConfig extends OAuth2IdentityProviderConfig {
         getConfig().put(ISSUER, issuer);
     }
     public String getLogoutUrl() {
-        return getConfig().get("logoutUrl");
+        return getConfig().get(LOGOUT_URL);
     }
     public void setLogoutUrl(String url) {
-        getConfig().put("logoutUrl", url);
+        getConfig().put(LOGOUT_URL, url);
     }
 
     public boolean isSendClientIdOnLogout() {