+++ services/src/main/java/org/keycloak/utils/OpenIdFederationTrustChainProcessor.java
@@ -1,8 +1,11 @@
-package org.keycloak.protocol.oidc.federation;
+package org.keycloak.utils;
 
 import jakarta.ws.rs.core.Response;
 import org.jboss.logging.Logger;
+import org.keycloak.TokenCategory;
 import org.keycloak.common.VerificationException;
+import org.keycloak.crypto.KeyType;
+import org.keycloak.crypto.KeyUse;
 import org.keycloak.crypto.PublicKeysWrapper;
 import org.keycloak.crypto.RS256SignatureProviderFactory;
 import org.keycloak.crypto.SignatureProvider;
@@ -12,12 +15,14 @@ import org.keycloak.exceptions.MetadataPolicyException;
 import org.keycloak.jose.JOSEParser;
 import org.keycloak.jose.jwk.JSONWebKeySet;
 import org.keycloak.jose.jwk.JWK;
+import org.keycloak.jose.jwk.JWKBuilder;
 import org.keycloak.jose.jws.JWSInput;
 import org.keycloak.jose.jws.JWSInputException;
 import org.keycloak.models.KeycloakSession;
+import org.keycloak.protocol.oidc.federation.MetadataPolicyUtils;
 import org.keycloak.representations.openid_federation.EntityStatement;
 import org.keycloak.representations.openid_federation.RPMetadataPolicy;
-import org.keycloak.representations.openid_federation.TrustChainForExplicit;
+import org.keycloak.representations.openid_federation.TrustChainResolution;
 import org.keycloak.services.ErrorResponseException;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -25,6 +30,7 @@ import java.time.LocalDateTime;
 import java.time.ZoneOffset;
 import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
@@ -42,62 +48,49 @@ public class OpenIdFederationTrustChainProcessor implements TrustChainProcessor 
     }
 
     /**
-     * This should construct all possible trust chains from a given leaf node url to a set of trust anchor urls
-     * @param leafNodeBaseUrl  this url should point to the base path of the leaf node (without the .well-known discovery subpath)
-     * @param trustAnchorIds this should hold the trust anchor ids
-     * @return any valid trust chains from the leaf node to the trust anchor.
-     * @throws IOException
-     */
-    public List<TrustChainForExplicit> constructTrustChainsFromUrl(String leafNodeBaseUrl, Set<String> trustAnchorIds, boolean policyRequired) throws IOException, InvalidTrustChainException {
-        String encodedLeafES = OpenIdFederationUtils.getSelfSignedToken(leafNodeBaseUrl, session);
-        EntityStatement statement = parseAndValidateSelfSigned(encodedLeafES);
-        return constructTrustChains(statement, trustAnchorIds, policyRequired);
-    }
-
-
-    /**
      * This should construct all possible trust chains from a given leaf node self-signed and encoded JWT to a set of trust anchor urls
      * @param leafEs  this is the EntityStatement of a leaf node (Relay party or Openid Provider)
      * @param trustAnchorIds this should hold the trust anchor ids
      * @return any valid trust chains from the leaf node JWT to the trust anchor.
      */
     @Override
-    public List<TrustChainForExplicit> constructTrustChains(EntityStatement leafEs, Set<String> trustAnchorIds, boolean policyRequired) {
+    public List<TrustChainResolution> constructTrustChains(EntityStatement leafEs, Set<String> trustAnchorIds, boolean policyRequired, boolean forRp) {
 
-        List<TrustChainForExplicit> trustChainForExplicits = subTrustChains(leafEs, trustAnchorIds, new HashSet<>());
+        List<TrustChainResolution> trustChainResolutions = subTrustChains(leafEs, leafEs, trustAnchorIds, new HashSet<>(), forRp);
 
-        return trustChainForExplicits.stream().map(trustChainForExplicit -> {
+        return trustChainResolutions.stream().map(trustChainResolution -> {
                     //combine policies if valid till now
-                    List<EntityStatement> parsedChain = trustChainForExplicit.getParsedChain();
-                    if (trustChainForExplicit != null && policyRequired) {
-                        RPMetadataPolicy combinedPolicy = parsedChain.get(parsedChain.size() - 1).getMetadataPolicy().getRelyingPartyMetadataPolicy();
-                        for (int i = parsedChain.size() - 2; i > 0; i--) {
-                            try {
+                    List<EntityStatement> parsedChain = trustChainResolution.getParsedChain();
+                    if (trustChainResolution != null && policyRequired) {
+                        try {
+                            RPMetadataPolicy combinedPolicy = parsedChain.get(parsedChain.size() - 1).getMetadataPolicy() == null ? null : parsedChain.get(parsedChain.size() - 1).getMetadataPolicy().getRelyingPartyMetadataPolicy();
+                            for (int i = parsedChain.size() - 2; i > 0; i--) {
                                 combinedPolicy = MetadataPolicyUtils.combineClientPolicies(combinedPolicy, parsedChain.get(i).getMetadataPolicy().getRelyingPartyMetadataPolicy());
-                            } catch (MetadataPolicyCombinationException e) {
-                                logger.debug(String.format("Cannot combine metadata policy of iss=%s sub=%s and its inferiors", parsedChain.get(i).getIssuer(), parsedChain.get(i).getSubject()));
-                                combinedPolicy = null;
                             }
+
+                            trustChainResolution.setCombinedPolicy(combinedPolicy);
+                            trustChainResolution.setTrustAnchorId(trustChainResolution.getParsedChain().get(trustChainResolution.getParsedChain().size() - 1).getIssuer());
+                            trustChainResolution.setLeafId(trustChainResolution.getParsedChain().get(0).getIssuer());
+                        } catch (MetadataPolicyCombinationException e) {
+                            logger.warn(String.format("Cannot combine metadata policy"));
+                            trustChainResolution = null;
                         }
-                        if (combinedPolicy != null) {
-                            trustChainForExplicit.setCombinedPolicy(combinedPolicy);
-                            trustChainForExplicit.setTrustAnchorId(trustChainForExplicit.getParsedChain().get(trustChainForExplicit.getParsedChain().size() - 1).getIssuer());
-                            trustChainForExplicit.setLeafId(trustChainForExplicit.getParsedChain().get(0).getIssuer());
-                        } else {
-                            trustChainForExplicit = null;
-                        }
+
+                    } else if (trustChainResolution != null) {
+                        trustChainResolution.setTrustAnchorId(trustChainResolution.getParsedChain().get(trustChainResolution.getParsedChain().size() - 1).getIssuer());
+                        trustChainResolution.setLeafId(trustChainResolution.getParsedChain().get(0).getIssuer());
                     }
 
-                    return trustChainForExplicit;
+                    return trustChainResolution;
                 })
                 .filter(Objects::nonNull)
                 .collect(Collectors.toList());
 
     }
 
-    private List<TrustChainForExplicit> subTrustChains(EntityStatement leafEs, Set<String> trustAnchorIds, Set<String> visitedNodes) {
+    private List<TrustChainResolution> subTrustChains(EntityStatement initialEntity, EntityStatement leafEs, Set<String> trustAnchorIds, Set<String> visitedNodes, boolean forRp) {
 
-        List<TrustChainForExplicit> chainsList = new ArrayList<>();
+        List<TrustChainResolution> chainsList = new ArrayList<>();
         visitedNodes.add(leafEs.getIssuer());
 
         if (leafEs.getAuthorityHints() != null && !leafEs.getAuthorityHints().isEmpty()) {
@@ -114,7 +107,7 @@ public class OpenIdFederationTrustChainProcessor implements TrustChainProcessor 
 
                     String fedApiUrl = subNodeSelfES.getMetadata().getFederationEntity().getFederationFetchEndpoint();
                     String encodedSubNodeSubordinate = OpenIdFederationUtils.getSubordinateToken(fedApiUrl, leafEs.getIssuer(), session);
-                    EntityStatement subNodeSubordinateES = parseAndValidateSelfSigned(encodedSubNodeSubordinate, subNodeSelfES.getJwks());
+                    EntityStatement subNodeSubordinateES = parseAndValidateSelfSigned(encodedSubNodeSubordinate, EntityStatement.class, subNodeSelfES.getJwks());
                     //fetch endpoint contains jwks of leafEs. So, validate based on subNodeSelfES.
                     if (!validateEntityStatementFields(subNodeSubordinateES, authHint, leafEs.getIssuer())) {
                         throw new ErrorResponseException(Errors.INVALID_TRUST_CHAIN, "Trust chain is not valid", Response.Status.BAD_REQUEST);
@@ -122,12 +115,25 @@ public class OpenIdFederationTrustChainProcessor implements TrustChainProcessor 
                     logger.debug(String.format("EntityStatement of %s about %s. AuthHints: %s", subNodeSubordinateES.getIssuer(), subNodeSubordinateES.getSubject(), subNodeSubordinateES.getAuthorityHints()));
                     visitedNodes.add(subNodeSelfES.getIssuer());
                     if (trustAnchorIds.contains(authHint)) {
-                        TrustChainForExplicit trustAnchor = new TrustChainForExplicit();
+                        TrustChainResolution trustAnchor = new TrustChainResolution();
                         trustAnchor.getParsedChain().add(0, subNodeSelfES);
+                        if (initialEntity.getSubject().equals(subNodeSubordinateES.getSubject())) {
+                            //set initial entity if no intermediates entities between trust achor and initial entity
+                            if ((forRp && subNodeSubordinateES.getMetadata().getRelyingPartyMetadata() == null) || (!forRp && subNodeSubordinateES.getMetadata().getOpenIdProviderMetadata() == null ))
+                                throw new ErrorResponseException(Errors.INVALID_TRUST_CHAIN, "Trust chain is not valid", Response.Status.BAD_REQUEST);
+                            trustAnchor.setInitialEntity(subNodeSubordinateES);
+                        } else {
+                            String initialEntityFetch = OpenIdFederationUtils.getSubordinateToken(fedApiUrl, initialEntity.getIssuer(), session);
+                            EntityStatement initialEntityFetchStatement = parseAndValidateSelfSigned(initialEntityFetch, EntityStatement.class, subNodeSelfES.getJwks());
+                            if (!validateEntityStatementFields(subNodeSubordinateES, authHint, leafEs.getIssuer()) || (forRp && initialEntityFetchStatement.getMetadata().getRelyingPartyMetadata() == null) || (!forRp && initialEntityFetchStatement.getMetadata().getOpenIdProviderMetadata() == null )) {
+                                throw new ErrorResponseException(Errors.INVALID_TRUST_CHAIN, "Trust chain is not valid", Response.Status.BAD_REQUEST);
+                            }
+                            trustAnchor.setInitialEntity(initialEntityFetchStatement);
+                        }
                         chainsList.add(trustAnchor);
                     } else {
-                        List<TrustChainForExplicit> subList = subTrustChains(subNodeSelfES, trustAnchorIds, visitedNodes);
-                        for (TrustChainForExplicit tcr : subList) {
+                        List<TrustChainResolution> subList = subTrustChains(initialEntity, subNodeSelfES, trustAnchorIds, visitedNodes, forRp);
+                        for (TrustChainResolution tcr : subList) {
                             tcr.getParsedChain().add(0, subNodeSelfES);
                             chainsList.add(tcr);
                         }
@@ -139,7 +145,7 @@ public class OpenIdFederationTrustChainProcessor implements TrustChainProcessor 
             });
 
         } else if (trustAnchorIds.contains(leafEs.getIssuer())) {
-            TrustChainForExplicit trustAnchor = new TrustChainForExplicit();
+            TrustChainResolution trustAnchor = new TrustChainResolution();
             trustAnchor.getParsedChain().add(0, leafEs);
             chainsList.add(trustAnchor);
         }
@@ -149,9 +155,9 @@ public class OpenIdFederationTrustChainProcessor implements TrustChainProcessor 
     }
 
     @Override
-    public EntityStatement parseAndValidateSelfSigned (String token, JSONWebKeySet publicKey) throws IOException, JWSInputException, VerificationException {
+    public <T extends EntityStatement> T parseAndValidateSelfSigned (String token, Class<T> clazz, JSONWebKeySet publicKey) throws IOException, JWSInputException, VerificationException {
         JWSInput jws = (JWSInput)JOSEParser.parse(token);
-        EntityStatement statement = jws.readJsonContent(EntityStatement.class);
+        T statement = jws.readJsonContent(clazz);
         String signatureAlgorithm = jws.getHeader().getAlgorithm().name();
         SignatureProvider signatureProvider = session.getProvider(SignatureProvider.class, signatureAlgorithm);
 
@@ -192,11 +198,15 @@ public class OpenIdFederationTrustChainProcessor implements TrustChainProcessor 
 
     }
 
+    public boolean validateEntityStatementFields(EntityStatement statement, String issuer, String subject) {
+        return statement.getIssuer() == null || statement.getIssuer().equals(issuer) || statement.getSubject() == null || statement.getSubject().equals(subject) || statement.getIat() == null || LocalDateTime.now().toEpochSecond(ZoneOffset.UTC) > statement.getIat() || statement.getExp() == null || LocalDateTime.now().toEpochSecond(ZoneOffset.UTC) < statement.getExp();
+    }
+
     @Override
-    public TrustChainForExplicit findAcceptableMetadataPolicyChain(List<TrustChainForExplicit> trustChainForExplicits, EntityStatement statement) {
-        TrustChainForExplicit validChain = null;
+    public TrustChainResolution findAcceptableMetadataPolicyChain(List<TrustChainResolution> trustChainResolutions, EntityStatement statement) {
+        TrustChainResolution validChain = null;
         EntityStatement current = statement;
-        for (TrustChainForExplicit chain : trustChainForExplicits) {
+        for (TrustChainResolution chain : trustChainResolutions) {
             try {
                 current = MetadataPolicyUtils.applyPoliciesToRPStatement(current, chain.getCombinedPolicy());
                 validChain = chain;
@@ -208,8 +218,26 @@ public class OpenIdFederationTrustChainProcessor implements TrustChainProcessor 
         return validChain;
     }
 
-    private boolean validateEntityStatementFields(EntityStatement statement, String issuer, String subject) {
-        return statement.getIssuer() == null || statement.getIssuer().equals(issuer) || statement.getSubject() == null || statement.getSubject().equals(subject) || statement.getIat() == null || LocalDateTime.now().toEpochSecond(ZoneOffset.UTC) > statement.getIat() || statement.getExp() == null || LocalDateTime.now().toEpochSecond(ZoneOffset.UTC) < statement.getExp();
+    @Override
+    public JSONWebKeySet getKeySet() {
+        List<JWK> keys = new LinkedList<>();
+        session.keys().getKeysStream(session.getContext().getRealm())
+                .filter(k -> k.getStatus().isEnabled() && k.getUse().equals(KeyUse.SIG) && k.getPublicKey() != null && k.getAlgorithm().equals(session.tokens().signatureAlgorithm(TokenCategory.ENTITY_STATEMENT)))
+                .forEach(k -> {
+                    JWKBuilder b = JWKBuilder.create().kid(k.getKid()).algorithm(k.getAlgorithm());
+                    if (k.getType().equals(KeyType.RSA)) {
+                        keys.add(b.rsa(k.getPublicKey(), k.getCertificate()));
+                    } else if (k.getType().equals(KeyType.EC)) {
+                        keys.add(b.ec(k.getPublicKey()));
+                    }
+                });
+
+        JSONWebKeySet keySet = new JSONWebKeySet();
+
+        JWK[] k = new JWK[keys.size()];
+        k = keys.toArray(k);
+        keySet.setKeys(k);
+        return keySet;
     }
 
     @Override
@@ -219,8 +247,14 @@ public class OpenIdFederationTrustChainProcessor implements TrustChainProcessor 
 
     //nimbus implementation - to be removed
 //    public EntityStatement parseAndValidateSelfSigned(String token) throws InvalidTrustChainException {
-//        EntityStatement statement = parse(token);
+//        EntityStatement statement = parse(token, EntityStatement.class);
 //        validateToken(token, statement.getJwks());
+//        return statement;
+//    }
+//
+//    public <T extends EntityStatement> T parseAndValidateSelfSigned(String token, Class<T> clazz, JSONWebKeySet jwks) throws InvalidTrustChainException {
+//        T statement = parse(token, clazz);
+//        validateToken(token, jwks);
 //        return statement;
 //    }
 //
@@ -272,22 +306,23 @@ public class OpenIdFederationTrustChainProcessor implements TrustChainProcessor 
 //
 //        JWSKeySelector<SecurityContext> keySelector = new JWSVerificationKeySelector<>(algs, keySource);
 //        jwtProcessor.setJWSKeySelector(keySelector);
-//        jwtProcessor.setJWSTypeVerifier(new DefaultJOSEObjectTypeVerifier<>(Stream.of(new JOSEObjectType("entity-statement+jwt")).collect(Collectors.toSet())));
+//        jwtProcessor.setJWSTypeVerifier(new DefaultJOSEObjectTypeVerifier<>(Stream.of(new JOSEObjectType(TokenUtil.ENTITY_STATEMENT_JWT), new JOSEObjectType(TokenUtil.EXPLICIT_REGISTRATION_RESPONSE_JWT)).collect(Collectors.toSet())));
 //        return jwtProcessor;
 //    }
 //
 //
 //
-//    public EntityStatement parse(String token) throws InvalidTrustChainException {
-//        String [] splits = token.split("\\.");
-//        if(splits.length != 3)
+//    public <T extends EntityStatement> T parse(String token, Class<T> clazz) throws InvalidTrustChainException {
+//        String[] splits = token.split("\\.");
+//        if (splits.length != 3)
 //            throw new InvalidTrustChainException("Trust chain contains a chain-link which does not abide to the dot-delimited format of xxx.yyy.zzz");
 //        try {
-//            return om.readValue(Base64.getDecoder().decode(splits[1]), EntityStatement.class);
+//            return om.readValue(Base64.getDecoder().decode(splits[1]), clazz);
 //        } catch (IOException e) {
 //            throw new InvalidTrustChainException("Trust chain does not contain a valid Entity Statement");
 //        }
 //    }
+
 
 
 }