+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/oidc/OpenIdFederationWellKnownProviderTest.java
@@ -6,20 +6,25 @@ import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.UriBuilder;
 import org.junit.Ignore;
 import org.junit.Test;
+import org.keycloak.OAuth2Constants;
 import org.keycloak.admin.client.resource.RealmResource;
 import org.keycloak.protocol.oidc.federation.OpenIdFederationWellKnownProviderFactory;
 import org.keycloak.representations.idm.OpenIdFederationRepresentation;
 import org.keycloak.representations.idm.RealmRepresentation;
 import org.keycloak.representations.openid_federation.EntityStatement;
 import org.keycloak.representations.openid_federation.OPMetadata;
+import org.keycloak.representations.openid_federation.RPMetadata;
+import org.keycloak.services.resources.IdentityBrokerService;
 import org.keycloak.services.resources.RealmsResource;
 import org.keycloak.testsuite.Assert;
 import org.keycloak.testsuite.util.AdminClientUtil;
 import org.keycloak.testsuite.util.oauth.OAuthClient;
 import org.keycloak.util.TokenUtil;
+import org.keycloak.utils.OpenIdFederationUtils;
 
 import java.io.IOException;
 import java.net.URI;
+import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -47,7 +52,8 @@ public class OpenIdFederationWellKnownProviderTest extends AbstractWellKnownProv
             OpenIdFederationRepresentation openIdFederationRepresentation = new OpenIdFederationRepresentation();
             openIdFederationRepresentation.setTrustAnchor("https://edugain.org/trust-anchor");
             openIdFederationRepresentation.setClientRegistrationTypesSupported(Stream.of("EXPLICIT").collect(Collectors.toList()));
-            openIdFederationRepresentation.setEntityTypes(Stream.of("OPENID_PROVIDER").collect(Collectors.toList()));
+            openIdFederationRepresentation.setEntityTypes(Stream.of("OPENID_PROVIDER", "OPENID_RELYING_PARTY").collect(Collectors.toList()));
+            Map<String, String> idpConfiguration = Map.of(OpenIdFederationUtils.SUBJECT_TYPES_SUPPORTED, "public");
             testRealm.openIdFederationsResource().create(openIdFederationRepresentation);
 
             //When Open Id Federation is configured
@@ -65,6 +71,16 @@ public class OpenIdFederationWellKnownProviderTest extends AbstractWellKnownProv
             assertEquals("explicit", op.getClientRegistrationTypes().get(0));
             assertEquals(UriBuilder.fromUri(OAuthClient.AUTH_SERVER_ROOT).path(RealmsResource.class).path(RealmsResource.class, "getOpenIdFederationClientsService").build("test").toString(), op.getFederationRegistrationEndpoint());
             testOidc(op);
+
+            RPMetadata rp = statement.getMetadata().getRelyingPartyMetadata();
+            Assert.assertNotNull("RPMetadata can not be null", rp);
+            assertEquals(1, rp.getClientRegistrationTypes().size());
+            assertEquals("explicit", rp.getClientRegistrationTypes().get(0));
+            assertEquals("web", rp.getApplicationType());
+            assertEquals(1, rp.getGrantTypes().size());
+            assertEquals(OAuth2Constants.AUTHORIZATION_CODE, rp.getGrantTypes().get(0));
+            assertEquals(1, rp.getRedirectUris().size());
+            assertEquals(IdentityBrokerService.openidFederationRedirectUri(UriBuilder.fromUri(OAuthClient.AUTH_SERVER_ROOT)).build("test").toString(), rp.getRedirectUris().get(0));
 
             realmRep.setOpenIdFederationEnabled(false);
             testRealm.update(realmRep);